Adapt `trace_filter`
