+++ client/rpc/trace/src/lib.rs
@@ -55,7 +55,7 @@ use fp_rpc::EthereumRuntimeRPCApi;
 pub use moonbeam_rpc_core_trace::{
 	FilterRequest, RequestBlockId, RequestBlockTag, Trace as TraceT, TraceServer, TransactionTrace,
 };
-use moonbeam_rpc_primitives_debug::{block, proxy, DebugRuntimeApi};
+use moonbeam_rpc_primitives_debug::{block, proxy, DebugRuntimeApi, V2_RUNTIME_VERSION};
 
 /// RPC handler. Will communicate with a `CacheTask` through a `CacheRequester`.
 pub struct Trace<B, C> {
@@ -798,6 +798,10 @@ where
 		let height = *block_header.number();
 		let substrate_parent_id = BlockId::<B>::Hash(*block_header.parent_hash());
 
+		// Runtime version
+		let runtime_version = api
+			.version(&substrate_parent_id)
+			.map_err(|e| internal_err(format!("Runtime api access error: {:?}", e)))?;
 		// Get `DebugRuntimeApi` version.
 		let api_version = api
 			.api_version::<dyn DebugRuntimeApi<B>>(&substrate_parent_id)
@@ -850,94 +854,157 @@ where
 
 		// Trace the block.
 		let f = || {
-			if api_version >= 2 {
-				let _result = api
-					.trace_block(&substrate_parent_id, &block_header, extrinsics)
-					.map_err(|e| {
-						internal_err(format!(
-							"Blockchain error when replaying block {} : {:?}",
-							height, e
-						))
-					})?
-					.map_err(|e| {
-						tracing::warn!(
-							"Internal runtime error when replaying block {} : {:?}",
-							height,
-							e
-						);
-						internal_err(format!(
-							"Internal runtime error when replaying block {} : {:?}",
-							height, e
-						))
-					})?;
-				Ok(proxy::v1::Result::V2(proxy::v1::ResultV2::Block))
-			} else {
-				// For versions < 2 block needs to be manually initialized.
-				api.initialize_block(&substrate_parent_id, &block_header)
-					.map_err(|e| internal_err(format!("Runtime api access error: {:?}", e)))?;
+			match runtime_version.spec_version {
+				version if version >= V2_RUNTIME_VERSION => {
+					let _result = api
+						.trace_block(&substrate_parent_id, &block_header, extrinsics)
+						.map_err(|e| {
+							internal_err(format!(
+								"Blockchain error when replaying block {} : {:?}",
+								height, e
+							))
+						})?
+						.map_err(|e| {
+							tracing::warn!(
+								"Internal runtime error when replaying block {} : {:?}",
+								height,
+								e
+							);
+							internal_err(format!(
+								"Internal runtime error when replaying block {} : {:?}",
+								height, e
+							))
+						})?;
+					Ok(proxy::v1::Result::V2(proxy::v1::ResultV2::Block))
+				}
+				_ => {
+					if api_version >= 2 {
+						let _result = api
+							.trace_block(&substrate_parent_id, &block_header, extrinsics)
+							.map_err(|e| {
+								internal_err(format!(
+									"Blockchain error when replaying block {} : {:?}",
+									height, e
+								))
+							})?
+							.map_err(|e| {
+								tracing::warn!(
+									"Internal runtime error when replaying block {} : {:?}",
+									height,
+									e
+								);
+								internal_err(format!(
+									"Internal runtime error when replaying block {} : {:?}",
+									height, e
+								))
+							})?;
+						Ok(proxy::v1::Result::V2(proxy::v1::ResultV2::Block))
+					} else {
+						// For versions < 2 block needs to be manually initialized.
+						api.initialize_block(&substrate_parent_id, &block_header)
+							.map_err(|e| {
+								internal_err(format!("Runtime api access error: {:?}", e))
+							})?;
 
-				#[allow(deprecated)]
-				let result = api.trace_block_before_version_2(&substrate_parent_id, extrinsics)
-					.map_err(|e| {
-						internal_err(format!(
-							"Blockchain error when replaying block {} : {:?}",
-							height, e
-						))
-					})?
-					.map_err(|e| {
-						tracing::warn!(
-							"Internal runtime error when replaying block {} : {:?}",
-							height,
-							e
-						);
-						internal_err(format!(
-							"Internal runtime error when replaying block {} : {:?}",
-							height, e
-						))
-					})?;
-				Ok(proxy::v1::Result::V1(proxy::v1::ResultV1::Block(result)))
-			}
-		};
-
-		let mut proxy = proxy::v1::CallListProxy::new();
-
-		if api_version >= 2 {
-			proxy.using(f)?;
-			let mut traces: Vec<_> = proxy.into_tx_traces();
-			// Fill missing data.
-			for trace in traces.iter_mut() {
-				trace.block_hash = eth_block_hash;
-				trace.block_number = height;
-				trace.transaction_hash = eth_transactions
-					.get(trace.transaction_position as usize)
-					.ok_or_else(|| {
-						tracing::warn!(
-							"Bug: A transaction has been replayed while it shouldn't (in block {}).",
-							height
-						);
-
-						internal_err(format!(
-							"Bug: A transaction has been replayed while it shouldn't (in block {}).",
-							height
-						))
-					})?
-					.transaction_hash;
-
-				// Reformat error messages.
-				if let block::TransactionTraceOutput::Error(ref mut error) = trace.output {
-					if error.as_slice() == b"execution reverted" {
-						*error = b"Reverted".to_vec();
+						#[allow(deprecated)]
+						let result = api.trace_block_before_version_2(&substrate_parent_id, extrinsics)
+							.map_err(|e| {
+								internal_err(format!(
+									"Blockchain error when replaying block {} : {:?}",
+									height, e
+								))
+							})?
+							.map_err(|e| {
+								tracing::warn!(
+									"Internal runtime error when replaying block {} : {:?}",
+									height,
+									e
+								);
+								internal_err(format!(
+									"Internal runtime error when replaying block {} : {:?}",
+									height, e
+								))
+							})?;
+						Ok(proxy::v1::Result::V1(proxy::v1::ResultV1::Block(result)))
 					}
 				}
 			}
-			Ok(traces)
-		} else {
-			match proxy.using(f) {
-				Ok(proxy::v1::Result::V1(proxy::v1::ResultV1::Block(result))) => Ok(result),
-				Err(e) => Err(e),
-				_ => Err(internal_err(format!(
-					"Bug: Api and result versions must match"
-				))),
+		};
+
+		match runtime_version.spec_version {
+			version if version >= V2_RUNTIME_VERSION => {
+				let mut proxy = proxy::v2::call_list::Listener::default();
+				proxy.using(f)?;
+				let mut traces: Vec<_> = proxy.into_tx_traces();
+				// Fill missing data.
+				for trace in traces.iter_mut() {
+					trace.block_hash = eth_block_hash;
+					trace.block_number = height;
+					trace.transaction_hash = eth_transactions
+						.get(trace.transaction_position as usize)
+						.ok_or_else(|| {
+							tracing::warn!(
+								"Bug: A transaction has been replayed while it shouldn't (in block {}).",
+								height
+							);
+
+							internal_err(format!(
+								"Bug: A transaction has been replayed while it shouldn't (in block {}).",
+								height
+							))
+						})?
+						.transaction_hash;
+
+					// Reformat error messages.
+					if let block::TransactionTraceOutput::Error(ref mut error) = trace.output {
+						if error.as_slice() == b"execution reverted" {
+							*error = b"Reverted".to_vec();
+						}
+					}
+				}
+				Ok(traces)
+			}
+			_ => {
+				let mut proxy = proxy::v1::CallListProxy::new();
+				if api_version >= 2 {
+					proxy.using(f)?;
+					let mut traces: Vec<_> = proxy.into_tx_traces();
+					// Fill missing data.
+					for trace in traces.iter_mut() {
+						trace.block_hash = eth_block_hash;
+						trace.block_number = height;
+						trace.transaction_hash = eth_transactions
+							.get(trace.transaction_position as usize)
+							.ok_or_else(|| {
+								tracing::warn!(
+									"Bug: A transaction has been replayed while it shouldn't (in block {}).",
+									height
+								);
+
+								internal_err(format!(
+									"Bug: A transaction has been replayed while it shouldn't (in block {}).",
+									height
+								))
+							})?
+							.transaction_hash;
+
+						// Reformat error messages.
+						if let block::TransactionTraceOutput::Error(ref mut error) = trace.output {
+							if error.as_slice() == b"execution reverted" {
+								*error = b"Reverted".to_vec();
+							}
+						}
+					}
+					Ok(traces)
+				} else {
+					match proxy.using(f) {
+						Ok(proxy::v1::Result::V1(proxy::v1::ResultV1::Block(result))) => Ok(result),
+						Err(e) => Err(e),
+						_ => Err(internal_err(format!(
+							"Bug: Api and result versions must match"
+						))),
+					}
+				}
 			}
 		}
 	}
